requirements.txt
fastapi
uvicorn
python-multipart
pillow
google-generativeai
geocoder
joblib
scikit-learn
pydantic
requests
python-dotenv